requirements.txt
aiogram==3.17.0
aiohttp==3.11.11
aiosqlite==0.20.0
fastapi==0.115.7
pydantic==2.10.6
pytest==8.3.4
python-dotenv==1.0.1
python_jose==3.3.0
Requests==2.32.3
